+++ lib/inquisition/config/locales/en.yml
@@ -10,8 +10,8 @@ en:
     config_already_exist: "Config for %{linter_name} already exists"
     run_all_linters: "Running all linters"
   errors:
-    config_absense_error: "Run inquisition build first"
-    no_gem_error: "Gem %{name} not found"
+    config_absense: "Run inquisition build first"
+    no_gem: "Gem %{name} not found"
     no_config_file: "File %{name} not found"
     invalid_ruby_version: "Invalid ruby version %{version}"
     absence_additional_software: "Additional software with command %{command} not found"
