+++ lib/inquisition/errors/base_config_absense_error.rb
@@ -2,7 +2,7 @@ module Inquisition
   module Errors
     class BaseConfigAbsenseError < BaseError
       def initialize
-        super(I18n.t('errors.config_absense_error'))
+        super(I18n.t('errors.config_absense'))
       end
     end
   end
