+++ lib/inquisition/errors/no_gem_error.rb
@@ -2,7 +2,7 @@ module Inquisition
   module Errors
     class NoGemError < BaseError
       def initialize(name)
-        super(I18n.t('errors.no_gem_error', name: name))
+        super(I18n.t('errors.no_gem', name: name))
       end
     end
   end
