+++ lib/inquisition/runner.rb
@@ -1,32 +1,20 @@
 module Inquisition
-  class Runner
-    extend Inquisition::Authorization
+  class Runner < Core::MultipleRunner
+    include Inquisition::Authorization
 
-    class << self
-      def call
-        authenticate_runner
-        run_runners
-      end
+    def call
+      authenticate_runner
+      super
+    end
 
-      private
+    private
 
-      def run_runners
-        runners.each do |runner, path|
-          runner.call if BaseConfig.config_enabled?(*path)
-        end
-      end
-
-      def runners
-        {
-          Auditors::Backend::Runner => config_path(:backend)
-          # Auditors::Frontend::Runner => config_path(:frontend)
-          # Auditors::Common::Runner => config_path(:common)
-        }
-      end
-
-      def config_path(*path)
-        [:linters, *path]
-      end
+    def runners
+      {
+        Auditors::Backend::Runner => config_path(:backend)
+        # Auditors::Frontend::Runner => config_path(:frontend)
+        # Auditors::Common::Runner => config_path(:common)
+      }
     end
   end
 end
